split signup
