+++ shared/constants/reducer.js
@@ -16,7 +16,7 @@ import type {State as Profile} from '../constants/types/profile'
 import type {State as Push} from '../constants/types/push'
 import type {State as RouteTree} from '../constants/types/route-tree'
 import type {State as Settings} from '../constants/types/settings'
-import type {State as Signup} from '../constants/signup'
+import type {State as Signup} from '../constants/types/signup'
 import type {State as Tracker} from '../constants/types/tracker'
 import type {State as UnlockFolders} from '../constants/unlock-folders'
 import type {State as Waiting} from '../constants/waiting'
