+++ shared/constants/signup.js
@@ -1,35 +1,6 @@
 // @flow
-import HiddenString from '../util/hidden-string'
+import type {State} from './types/signup'
 import {isMobile} from '../constants/platform'
-
-export type AutoInviteRequestState = 'NotRequested' | 'Requested' | 'Done'
-
-export type State = {
-  deviceName: ?string,
-  deviceNameError: ?string,
-  email: ?string,
-  emailError: ?Error,
-  inviteCode: ?string,
-  inviteCodeError: ?string,
-  nameError: ?Error,
-  paperkey: ?HiddenString,
-  passphrase: ?HiddenString,
-  passphraseError: ?HiddenString,
-  phase:
-    | 'inviteCode'
-    | 'usernameAndEmail'
-    | 'passphraseSignup'
-    | 'deviceName'
-    | 'signupLoading'
-    | 'success'
-    | 'signupError'
-    | 'requestInvite'
-    | 'requestInviteSuccess',
-  signupError: ?HiddenString,
-  username: ?string,
-  usernameError: ?Error,
-  waiting: boolean,
-}
 
 const initialState: State = {
   deviceName: isMobile ? 'Mobile Device' : 'Home Computer',
